Create Hook for KogitoChannelBUs
